enh: define size-like variables as size_t instead of integer so to avoid unnecessary type conversions
mnt: remove unused function

fix: converting the output of GetShapeProduct from size_t to int can lead to overflow, in this case a negative value

enh: improve GetMaxAllocAndUpperBoundBuffer and remove unnecessary or inadequate type conversions

enh: at GetShapeProduct, return SIZE_MAX in case of an integer overflow, and include defensive check for negative shape values

enh: consider the case where the result of GetShapeProduct is SIZE_MAX to avoid another overflow

enh: use datatypes as defined by TileAxis class (tiling_analyzer.h) and DimensionInfo struct (pass_info.h)

mnt: apply datatype changes on multicore_tiling module as well

mnt: make variable const reference and fix uneffective casting

enh: make Tensor shape_ int64_t for consistency

mnt: forgot to change a variable type

fix: revert changes on init_graph done by mistake

mnt: initialize c1_tiling_ with 1, boolean variable to check c1_tiling assignment

enh: rename tile_ as c0_tiling_, and better logic for assigning c1_tiling to dims at tiling.cc (thanks @dabaiji )

+++ src/poly/tiling/hermes/hardware.cc
@@ -21,8 +21,8 @@ namespace ir {
 namespace poly {
 size_t Hardware::mem_VC_alloc_failed_ = 0;
 
-Hardware::Hardware(int num_core, int mem_VC_size, int mem_C1_size, int mem_C0_size, int mem_VC_align, int mem_C1_align,
-                   int vblocknum, int vblocksize)
+Hardware::Hardware(size_t num_core, size_t mem_VC_size, size_t mem_C1_size, size_t mem_C0_size, size_t mem_VC_align,
+                   size_t mem_C1_align, size_t vblocknum, size_t vblocksize)
     : num_core_{num_core},
       mem_VC_size_{mem_VC_size / (1 << mem_VC_alloc_failed_)},  // we divide VC by 2 for each VC alloc error
       mem_C1_size_{mem_C1_size},
